FIX: [core] made more database functions postgres compatible
