+++ core/orgs/init.go
@@ -23,7 +23,8 @@ type OrgStore struct {
 
 type OrgAdapter interface {
 	Get(orgID string) (models.Org, error)
-
+	CheckOrgExists() (orgID string, err error)
+	CreateOrg(org *models.Org) error
 	GetIDP(orgID, idpName string) (models.IdentityProvider, error)
 
 	RemoveAllManagedAccounts(orgID string) error
